requirements.txt
Flask
celery
slack_sdk
python-dotenv
motor
pymongo